Updated docu, changelog and fixxes for eclipse command
I updated the docu file and the changelog.

Sorry that I removed the mirror feature before, I readded it. I haven't noticed before that this was a feature itself...
I moved some doError Messages to the according function where they should actually be displayed.

I hope that this time I completed all the aspects of the DOD