Fix version-prefix matching issue
This is my final understanding of this issue: version-prefix is used to find the latest version STARTING with version-prefix. It should be noted that when e.g. version-prefix is "1.1*", only version numbers starting with "1.1.xxx" can be matched, while versions like "1.10.xxx" should not be matched. The version "1.10.xxx" is not the version that can be matched by "1.1*". To match "1.10.xxx", the version_prefix should be "1.10*"/"1.10.*"